Moved repository to eva-server for easier development. Mongo config file.
