+++ eva-server/src/main/java/uk/ac/ebi/eva/server/MyMongoConfig.java
@@ -1,29 +1,24 @@
-package uk.ac.ebi.eva.lib.repository;
+package uk.ac.ebi.eva.server;
 
-import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.context.annotation.Bean;
+import org.springframework.context.annotation.Configuration;
 import org.springframework.core.convert.converter.Converter;
-import org.springframework.data.mongodb.MongoDbFactory;
-import org.springframework.data.mongodb.core.MongoTemplate;
 import org.springframework.data.mongodb.core.convert.CustomConversions;
 import org.springframework.data.mongodb.core.convert.DbRefResolver;
 import org.springframework.data.mongodb.core.convert.DefaultDbRefResolver;
 import org.springframework.data.mongodb.core.convert.MappingMongoConverter;
 import org.springframework.data.mongodb.core.mapping.MongoMappingContext;
-import org.springframework.stereotype.Repository;
 import uk.ac.ebi.eva.commons.models.converters.data.MongoDBObjectToVariantConverter;
 
+import java.io.IOException;
 import java.util.ArrayList;
 import java.util.List;
 
 /**
  * @author Tom Smith
  */
-@Repository
-public class VariantRepositoryImpl {
-
-    @Autowired
-    private MongoDbFactory mongoDbFactory;
+@Configuration
+public class MyMongoConfig extends ApplicationConfiguration {
 
     @Bean
     public CustomConversions customConversions() {
@@ -33,18 +28,13 @@ public class VariantRepositoryImpl {
     }
 
     @Bean
-    public MappingMongoConverter mongoConverter() {
+    public MappingMongoConverter mongoConverter() throws IOException {
         MongoMappingContext mappingContext = new MongoMappingContext();
-        DbRefResolver dbRefResolver = new DefaultDbRefResolver(mongoDbFactory);
+        DbRefResolver dbRefResolver = new DefaultDbRefResolver(mongoDbFactory());
         MappingMongoConverter mongoConverter = new MappingMongoConverter(dbRefResolver, mappingContext);
         mongoConverter.setCustomConversions(customConversions());
         mongoConverter.afterPropertiesSet();
         return mongoConverter;
-    }
-
-    @Bean
-    public MongoTemplate mongoTemplate() {
-        return new MongoTemplate(mongoDbFactory, mongoConverter());
     }
 
 }
